+++ .github/workflows/release_wheel.yml
@@ -43,14 +43,12 @@ jobs:
       - name: Build wheel
         run: |
           chown -R $CI_UID:$CI_GID "$GITHUB_WORKSPACE"
-          version="$(cat version.txt)"
           docker run --rm -t \
               -v "$CI_RUNNER_CACHE_DIR":/ci-cache \
               -v "$GITHUB_WORKSPACE":/app \
               -e FLASHINFER_CI_PYTHON_VERSION=${{ matrix.python }} \
               -e FLASHINFER_CI_CUDA_VERSION=${{ matrix.cuda }} \
               -e FLASHINFER_CI_TORCH_VERSION=${{ matrix.torch }} \
-              -e FLASHINFER_BUILD_VERSION=$version \
               -e TORCH_CUDA_ARCH_LIST="$TORCH_CUDA_ARCH_LIST" \
               -e MAX_JOBS=128 \
               --user $CI_UID:$CI_GID \
