--- python/aot_MANIFEST.in
@@ -1,12 +1,0 @@
-# MANIFEST.in for AOT wheel
-
-prune */__pycache__
-prune csrc
-exclude aot_setup.py
-exclude setup.py
-
-include flashinfer/data/version.txt
-graft flashinfer/data/csrc
-graft flashinfer/data/include
-graft flashinfer/data/cutlass/include
-graft flashinfer/data/cutlass/tools/util/include
