--- python/jit_MANIFEST.in
@@ -1,16 +1,0 @@
-# MANIFEST.in for JIT sdist
-
-global-exclude *.so
-
-prune */__pycache__
-prune csrc
-exclude flashinfer/jit/aot_config.py
-exclude aot_setup.py
-exclude mypy.ini
-exclude pylintrc
-
-include flashinfer/data/version.txt
-graft flashinfer/data/csrc
-graft flashinfer/data/include
-graft flashinfer/data/cutlass/include
-graft flashinfer/data/cutlass/tools/util/include
